changed dto design - dogsProfileDto

+++ demo/demo/src/main/java/com/example/demo/entity/Owner.java
@@ -16,6 +16,6 @@ public class Owner {
     Set<Dog> dogs;
     @Column
             // @Column 어노테이션이 없으면 이 컬럼 옵션이 nullable = false가 됨
-    Integer ownerPoints;
+    Long ownerPoints;
     // 숫자가 커질 것 같으면 Long으로 바꾸기
 }
